refactor(project): remove default dictionary includes
