--- data/common/index.js
@@ -1,6 +1,0 @@
-import passwords from "./passwords.json"
-
-export default {
-  userInput: [],
-  passwords
-}
