--- data/de/index.js
@@ -1,10 +1,0 @@
-import commonWords from "./commonWords.json"
-import firstnames from "./firstnames.json"
-import lastnames from "./lastnames.json"
-
-export default {
-  userInput: [],
-  commonWords,
-  firstnames,
-  lastnames
-}
