--- data/en/index.js
@@ -1,12 +1,0 @@
-import commonWords from "./commonWords.json"
-import firstnames from "./firstnames.json"
-import lastnames from "./lastnames.json"
-import wikipedia from "./wikipedia.json"
-
-export default {
-  userInput: [],
-  commonWords,
-  firstnames,
-  lastnames,
-  wikipedia
-}
